Refactored rawtransaction unit tests
